requirements.txt
art==6.5
gTTS==2.5.4
pygame==2.6.1
pyttsx3==2.99
PyYAML==6.0.2
sounddevice==0.5.2
transliterate==1.10.2
vosk==0.3.45